+++ packages/hardhat-deploy/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @midl-xyz/hardhat-deploy
 
+## 2.0.0-next.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-react@2.0.0-next.24
+  - @midl-xyz/midl-js-executor@2.0.0-next.24
+
+## 2.0.0-next.23
+
+### Patch Changes
+
+- Updated dependencies [5197f61]
+  - @midl-xyz/midl-js-executor@2.0.0-next.23
+  - @midl-xyz/midl-js-react@2.0.0-next.23
+  - @midl-xyz/midl-js-core@2.0.0-next.23
+  - @midl-xyz/midl-js-node@2.0.0-next.23
+
+## 2.0.0-next.22
+
+### Patch Changes
+
+- fix: etna to regtest
+- Updated dependencies
+  - @midl-xyz/midl-js-executor@2.0.0-next.22
+
 ## 2.0.0-next.21
 
 ### Patch Changes