+++ packages/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @midl-xyz/midl-js-react
 
+## 2.0.0-next.24
+
+### Patch Changes
+
+- feat(react): use balance default account
+
+## 2.0.0-next.23
+
+### Patch Changes
+
+- 5197f61: feat(core): define default account
+- Updated dependencies [5197f61]
+  - @midl-xyz/midl-js-core@2.0.0-next.23
+
 ## 2.0.0-next.18
 
 ### Patch Changes