+++ packages/satoshi-kit/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @midl-xyz/satoshi-kit
 
+## 2.0.0-next.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-react@2.0.0-next.24
+
+## 2.0.0-next.23
+
+### Patch Changes
+
+- Updated dependencies [5197f61]
+  - @midl-xyz/midl-js-react@2.0.0-next.23
+  - @midl-xyz/midl-js-core@2.0.0-next.23
+  - @midl-xyz/midl-js-connectors@2.0.0-next.23
+
 ## 2.0.0-next.20
 
 ### Patch Changes